+++ web/static/js/index.js
@@ -3,7 +3,7 @@ data = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20
   return Math.round(Math.random() * 100);
 });
 ml = new layout({
-  root: '.pd-layout'
+  root: '.pdl-layout'
 });
 ml.on('render', function(){
   var box, ref$;
@@ -19,7 +19,7 @@ ml.on('render', function(){
 ml.init(function(){
   var this$ = this;
   return (this.lc || (this.lc = {})).view = new ldView({
-    root: ld$.find('g.view', 0),
+    root: ld$.find('g[data-name=view]', 0),
     initRender: false,
     handler: {
       data: {