+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.0.3
+
+ - better scoping CSS class with `pdl` prefix. 
+ - unify class name and attribute between render and layout elements.
+ - add feature: auto generate SVG tree.
+
+
 # v0.0.2
 
  - use better namespaced class name.